perf(chat): extract ChatSearchOverlay into standalone View for narrow observation scope (LUM-1280)
When search is active (Cmd+F), searchMatches reads viewModel.messages
which registered an observation dependency in the outer ChatView.body
scope. Every streamed token invalidated the outer body and triggered
an O(n) filter recomputation on the main thread.

By extracting into a standalone View struct, the messages observation
dependency is scoped to ChatSearchOverlay.body only — the outer
ChatView.body is never invalidated by message changes during search.
